Massive performance optimization: Volume & Distribution sync with bulk operations
Major refactoring to eliminate N+1 queries and implement bulk operations across the codebase.

Performance improvements:
- Volume sync: Reduced from 300+ queries to 3-4 bulk queries (~1.2s total)
- Distribution sync: Reduced from 10,000+ queries to 5-10 bulk queries (~2.5s total)
- Added database-side ranking using ROW_NUMBER() window functions
- Implemented parallel processing with controlled concurrency (3 sports max)
- Smart change detection to minimize history records

Key changes:
- Reorganized SQLC structure: moved to pkg/database/generated/ to prevent deletion
- Created pkg/database/pool for connection pooling configuration
- Added 27 composite indexes for query optimization
- Implemented bulk operations using PostgreSQL UNNEST for array processing
- Flattened nested data structures for efficient processing
- Added comprehensive feature documentation for volume/distribution analytics

New features documented:
- Sharp vs Public Money Divergence Alerts
- Reverse Line Movement Detector
- Late Money Steam Tracker
- Public Fade System
- Smart Money Confidence Score

Technical improvements:
- Migrated from individual row operations to bulk upserts
- Preserved array ordering throughout bulk operations
- Added structured error handling and logging
- Optimized memory allocation with pre-calculated capacities

+++ pkg/models/api/response.go
@@ -56,9 +56,9 @@ type BigMoverResponse struct {
 	EventTime            time.Time `json:"event_time"`
 	EventStatus          string    `json:"event_status"`
 	IsLive               bool      `json:"is_live"`
-	HomeScore            *int32    `json:"home_score,omitempty"`
-	AwayScore            *int32    `json:"away_score,omitempty"`
-	MinuteOfMatch        *int32    `json:"minute_of_match,omitempty"`
+	HomeScore            *int64    `json:"home_score,omitempty"`
+	AwayScore            *int64    `json:"away_score,omitempty"`
+	MinuteOfMatch        *int64    `json:"minute_of_match,omitempty"`
 	BettingVolumePercent *float64  `json:"betting_volume_percent,omitempty"`
 	HomeTeamCountry      string    `json:"home_team_country"`
 	AwayTeamCountry      string    `json:"away_team_country"`
